Added extra information to output XML
The converted query string and the location of the query in the native
web application are now added as extra information (for every record) to
the output XML. Native URL still has to be escaped for the browser. And
the right native URLs for all the BlackLab corpora have to be added to
the config file.